Add mkiso.py and mkiso.d/.
mkiso.py generates a customized iso image that fully automated the installation
of a ubuntu 16 server OS with the following settings:
    locale: en_US
    user: geek
    pass: geek

For more details about the customized settings, please take a look at the
mkiso.d/geek.seed file.

The typical usage is:
    ./mkiso.py <iso-name>

You will be prompted for sudo password for mounting and unmounting iso images.
That is expected.