Add try_iso.py for more easily test the bootability of an iso.
The try_iso.py script creates a libvirt domain that boots from its cdrom, which
is connected with the iso image under test.  A typical usage is:
    ./try_iso.py <iso-filename>

You will see a virt-viewer window pops up.  If the iso image under test is
bootable, you can see that the virtual machine booted.

The boot order of the virtual is:
    1.  hard drive
    2.  cdrom

Initially the hard drive is empty, so it will boot from cdrom.  After the OS is
installed on the hard drive, it will boot from the hard drive so that you have a
chance of examining the system.

You can obviously run the above command as root.  With a little tweak, you can
also make it run without sudo.  Here is how, tested on Ubuntu 16.04:
    1.  Add the following line to /etc/libvirt/qemu.conf:
            dynamic_ownership = 0
    2.  Restart the libvirt-bin service:
            sudo service libvirt-bin restart

On Ubuntu 14.04, you probably also need to add the following line to
/etc/libvirt/qemu.conf:
    security_driver = "none"